abyss-drawgraph: Separate the two componenets
Set the origin of the layout. Pin the first contig and its reverse
complement to an extreme negative and positive position, so that the two
components should not overlap. The two vertices should be in separate
components. If the graph has multiple components, the matrix should be
regularized by adding lambda to all values on the diagonal. The
components will overlap be can be separated afterward.