BreadthFirstSearch.h: handle previously visited root node
Check if start node is already exhausted ("black") before initiating a
BFS. This allows multiple calls to breadthFirstSearch() on the same
graph with different starting points, while ensuring that areas that
overlap between searches are not visited more than once. The external
"color map" argument tracks the visited states of the nodes between
calls to breadthFirstSearch().