bloom-dbg: trim assembled k-mers from contig ends
Most of the time, reads are extended within mutually exclusive branches
of the de Bruijn graph and contigs do not overlap by more than k-1
bases. However there are edge cases where recurrent read errors and/or
coverage gaps cause multiple reads to be extended within the same branch
of the DBG, resulting in overlapping contigs. To address this, we trim
previously-assembled k-mers from each end of a contig before writing
them to the output FASTA.