konnector/sealer: max search cost option (-C)
This option specifies an upper bound on the number of edges that can be
traversed during each graph search. This fixes a long-standing issue with
extremely long runtimes for searches in dense areas of the graph.

The `-C`/`--max-cost` is intended to replace the `--max-branches` option, which
did not reliably bound graph searching time. I've changed the default value of
`--max-branches` to `nolimit` and moved it to the `Deprecated Options` section
of the `--help` output.