bloom-dbg: track start/end k-mers of short contigs in a hash set
If we check very short contigs (e.g. length = k) against the Bloom filter of
previously assembled k-mers (assembledKmerSet), we are likely to get full-length
matches purely due to Bloom filter false positives, which will cause the contig
to be incorrectly omitted from the output FASTA.

To address this issue, we track the start and end k-mers of short contigs in a
separate hash set.