bloom-dbg: extend from branching k-mers (major refactor)
instead of from read ends. We still extend from read ends when there are
branching k-mers detected within a read.  This ensures that genomic regions
which are "cut off" from neighbouring branch k-mers by coverage gaps
will not be missed.

This commit is a major refactoring and change in assembler behaviour.

I have noticed that this change really killed the performance, which may be
due to multithreaded contention for malloc.