ExtendPath.h: implement bidirectional look-ahead (`trueBranch`)
This is a major implementation change.

The core change is the replacement of the `lookAhead` function with the
`trueBranch` function. While `lookAhead` was a straight-forward unidirectional
depth-first search, `trueBranch` allows changes in a direction, provided that
the current branch is longer than the threshold for Bloom filter false positive
branches.

Other change this commit:

* introduction of the `fpTrim` parameter, indicating an expected upper bound
on the length of Bloom filter false positive branches
* introduction of the `successor` method, which returns the unambiguous
predecessor/successor of a vertex after trimming false branches (or an
error code if the successor is not unique)
* introduction of the `ambiguous` method, which return true if a vertex has
more than one possible predecessor/successor, after trimming false branches