bloom-dbg: change definition of tip
Change the definition of a tip, so that tips extending from the ends of bubbles
are not trimmed. (This more closely matches the behaviour of the MPI assembler.)

In the new definition of a tip, one end of the path must terminate in a dead
end, and the other end of the path must terminate in either a dead or a branch
k-mer having multiple branches _in the same direction_ as the tip. The previous
tip definition did not distinguish between branch k-mers with multiple incoming
branches vs. branch k-mers with multiple outgoing branches.

Implementing the change in tip definition required changing the result codes in
`ExtendPath.h` to distinguish between different types of branch vertices
(i.e. ambiguous-in vs. ambiguous-out). This in turn required some cascading
changes to konnector.h/bloom.cc.