BranchGroup.h: compile fix for clang-6 on Linux
Removed `std::swap` specialization for `BranchGroup`, which clang didn't like
because its `noexcept` clause mismatched the base implementation of `std::swap`.
The `BranchGroup` specialization was just a safety measure to prevent people
from incorrectly using `std::swap` on `BranchGroup`, so after trying to fix it
for a while, I just decided to remove it.