Fix FTBFS on 32-bit architectures
Version 1.5.2-1 of abyss FTBFS on 32-bit architectures (armhf, i386 and powerpc). Previous versions built fine on all architectures. . There were a couple of places in the code that was introduced in version 1.5.2 where size_t data type was used instead of a fixed size data type. The patch builds on amd64, i386 and armhf with and all unit tests passed.